requirements.txt
httpx
PyExecJS
loguru